requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==45.0.3
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
loguru==0.7.3
msgpack==1.1.0
multidict==6.4.3
nonebot-adapter-onebot==2.4.6
nonebot2==2.4.2
propcache==0.3.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pygtrie==2.5.0
PyMySQL==1.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.0
